backend/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-cors==4.0.0
flask-socketio==5.3.6
python-dotenv==1.0.0
psycopg2-binary==2.9.9
google-generativeai==0.3.2
python-socketio==5.10.0
eventlet==0.33.3
gunicorn==21.2.0
gevent==23.9.1
pytest==8.1.1
pytest-flask==1.3.0
pytest-cov==4.1.0 